--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - during grouping, group keys are now added automatically to the select list.
 - implement `sum()` reducer.
 - `==` operator now works for string columns too.
+- Improved performance of groupby operations.
 
 #### Fixed
 - fread will no longer emit an error if there is an NA string in the header.