--- a/c/expr/py_expr.cc
+++ b/c/expr/py_expr.cc
@@ -58,7 +58,7 @@ PyObject* expr_column(PyObject*, PyObject* args)
     PyErr_Format(PyExc_ValueError, "Invalid column index %lld", index);
   }
   Column* col = dt->columns[index]->shallowcopy(ri);
-  col->reify();
+  // col->reify();
   return pycolumn::from_column(col, nullptr, 0);
 }