--- a/c/py_columnset.cc
+++ b/c/py_columnset.cc
@@ -140,6 +140,7 @@ PyObject* columns_from_columns(PyObject*, PyObject* args)
       return nullptr;
     }
     reinterpret_cast<pycolumn::obj*>(elem)->ref = nullptr;
+    columns[i]->reify();
   }
   columns[ncols] = nullptr;