--- a/c/utils/array.h
+++ b/c/utils/array.h
@@ -48,10 +48,14 @@ template <typename T> class array
   private:
     T* x;
     size_t n;
+    bool owned;
+    int64_t : 56;
 
   public:
-    array(size_t len = 0) : x(nullptr), n(0) { resize(len); }
-    ~array() { std::free(x); }
+    array(size_t len = 0) : x(nullptr), n(0), owned(true) { resize(len); }
+    array(size_t len, const T* ptr)
+      : x(const_cast<T*>(ptr)), n(len), owned(false) {}
+    ~array() { if (owned) std::free(x); }
     // copy-constructor and assignment are forbidden
     array(const array<T>&) = delete;
     array<T>& operator=(const array<T>&) = delete;
@@ -64,12 +68,14 @@ template <typename T> class array
       using std::swap;
       swap(first.x, second.x);
       swap(first.n, second.n);
+      swap(first.owned, second.owned);
     }
 
     template <typename S> array<S> cast() {
       array<S> res;
       res.n = n * sizeof(T) / sizeof(S);
       res.x = reinterpret_cast<S*>(x);
+      res.owned = owned;
       x = nullptr;
       return res;
     }
@@ -79,7 +85,7 @@ template <typename T> class array
       size_t size = sizeof(T) * n;
       x = nullptr;
       n = 0;
-      return MemoryRange(size, ptr, /* own = */ true);
+      return MemoryRange(size, ptr, /* own = */ owned);
     }
 
     // Standard operators
@@ -95,6 +101,9 @@ template <typename T> class array
 
     void resize(size_t newn) {
       if (newn == n) return;
+      if (!owned) {
+        throw MemoryError() << "Cannot resize array: not owned";
+      }
       if (newn == 0) {
         std::free(x);
         x = nullptr;