* Tweaked `sort.max_radix_bits` constant to speed up sorting
* Groupby operations are now significantly faster (especially function `first`) because columns are no longer materialized before performing the reduction.

For the same operation as in #1095, we have now the following timings:
```
>>> y = x[:, {"v1": sum(f.v1)}, f.id1]
    call DataTable.rowindex
    done DataTable.rowindex in 0.000 ms
    call DataTable.stypes
    done DataTable.stypes in 0.008 ms
    call DataTable.rowindex
    done DataTable.rowindex in 0.000 ms
    call DataTable.sort(...)
    done DataTable.sort(...) in 85.602 ms
    call CLSNAME.expr_column(...)
    done CLSNAME.expr_column(...) in 0.006 ms
    call CLSNAME.expr_reduceop(...)
    done CLSNAME.expr_reduceop(...) in 0.078 ms
    call del Column
    done del Column in 0.005 ms
    call CLSNAME.expr_column(...)
    done CLSNAME.expr_column(...) in 0.003 ms
    call CLSNAME.expr_reduceop(...)
    done CLSNAME.expr_reduceop(...) in 69.257 ms
    call del Column
    done del Column in 0.001 ms
    call ColumnSet.columns_from_columns(...)
    done ColumnSet.columns_from_columns(...) in 0.018 ms
    call del Column
    done del Column in 0.000 ms
    call del Column
    done del Column in 0.000 ms
    call ColumnSet.to_datatable(...)
    done ColumnSet.to_datatable(...) in 0.001 ms
    call DataTable.nrows
    done DataTable.nrows in 0.001 ms
    call DataTable.ncols
    done DataTable.ncols in 0.000 ms
    call DataTable.nrows
    done DataTable.nrows in 0.000 ms
    call del ColumnSet
    done del ColumnSet in 0.000 ms
    call del Groupby
    done del Groupby in 0.003 ms
    call del RowIndex
    done del RowIndex in 0.000 ms
```

Thus,
* `sort()`: 221.671 ms -> 85.602 ms
* first col: 76.152 ms -> 0.084 ms
* second col: 179.940 ms -> 69.260 ms

Overall, the timings of 3 consecutive runs are: (171ms, 127ms, 123ms)
The same task in R data.table: (152ms, 153ms, 160ms)
The same task in pandas: (360ms, 320ms, 286ms)

Closes #1095